endrer navn til opprettKorreksjonsutkast som det også heter i frontend. endrer til integer på korreksjonsnummer. formatering Co-authored-by: Mats Bolstad <mats.bolstad@nav.no>

+++ src/main/resources/db/migration/V16__korreksjonsnummer.sql
@@ -1,1 +1,1 @@
-alter table refusjon add column korreksjonsnummer numeric;
+alter table refusjon add column korreksjonsnummer integer;
